# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Search Engine API
  version: "1.0"
  description: |
    A comprehensive content search engine API that aggregates content from multiple providers (JSON/XML),
    calculates relevance scores, and provides search and filtering capabilities.

    ## Features
    - Multi-provider content aggregation (JSON/XML)
    - Advanced scoring algorithm with freshness and engagement metrics
    - Full-text search with fuzzy matching
    - Content type filtering (video/text)
    - Rate limiting and caching
    - Admin dashboard for monitoring and management

    ## Authentication
    Admin endpoints require X-API-Key header with the configured API key.
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.search-engine.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/contents/search:
    get:
      summary: Search contents
      description: |
        Search through aggregated content with advanced filtering and sorting.

        Supports fuzzy search on titles and descriptions, content type filtering,
        and multiple sorting options including relevance scores.
      tags:
        - Content
      parameters:
        - name: q
          in: query
          description: Search keyword (supports fuzzy matching)
          required: false
          schema:
            type: string
            example: "python tutorial"
        - name: type
          in: query
          description: Content type filter
          required: false
          schema:
            type: string
            enum: [video, text]
            example: "video"
        - name: sort
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [score_desc, score_asc, date_desc, date_asc]
            default: score_desc
            example: "score_desc"
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: page_size
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/contents/{id}:
    get:
      summary: Get content details
      description: Retrieve detailed information about a specific content item
      tags:
        - Content
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID
          schema:
            type: integer
            format: int64
            example: 123
      responses:
        '200':
          description: Content details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentDetailResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/contents/{id}:
    delete:
      summary: Soft delete content
      description: Mark a content item as deleted (soft delete)
      tags:
        - Admin
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID to delete
          schema:
            type: integer
            format: int64
            example: 123
      responses:
        '200':
          description: Content deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                        example: 123
                      deleted_at:
                        type: string
                        format: date-time
        '400':
          description: Invalid content ID
        '404':
          description: Content not found
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key for authentication
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_PARAMETER"
            message:
              type: string
              example: "Invalid parameter value"
            details:
              type: object
              additionalProperties:
                type: string
    ContentSummaryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        title:
          type: string
          example: "Introduction to Go Programming"
        content_type:
          type: string
          enum: [video, text]
          example: "video"
        description:
          type: string
          nullable: true
          example: "Learn the basics of Go programming language"
        url:
          type: string
          nullable: true
          example: "https://example.com/video123"
        thumbnail_url:
          type: string
          nullable: true
          example: "https://example.com/thumbnail.jpg"
        score:
          type: number
          format: float
          example: 45.67
        published_at:
          type: string
          format: date-time
          nullable: true
        provider:
          type: string
          example: "provider1"
    PaginationDTO:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          format: int64
          example: 150
        total_pages:
          type: integer
          example: 8
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentSummaryDTO'
        pagination:
          $ref: '#/components/schemas/PaginationDTO'
        error:
          $ref: '#/components/schemas/ErrorResponse'
    MetricsDTO:
      type: object
      properties:
        views:
          type: integer
          format: int64
          nullable: true
          example: 15000
        likes:
          type: integer
          format: int64
          nullable: true
          example: 1200
        reading_time:
          type: integer
          nullable: true
          example: 15
        reactions:
          type: integer
          nullable: true
          example: 25
        recalculated_at:
          type: string
          format: date-time
          nullable: true
    ContentDetailDTO:
      allOf:
        - $ref: '#/components/schemas/ContentSummaryDTO'
        - type: object
          properties:
            metrics:
              $ref: '#/components/schemas/MetricsDTO'
    ContentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/ContentDetailDTO'
        error:
          $ref: '#/components/schemas/ErrorResponse'
...